requirements.txt
numpy
pandas
joblib
requests
ta
python-binance
binance-futures-connector
